Move index.js into src.
