+++ package.json
@@ -2,6 +2,7 @@
   "name": "axdd-components",
   "version": "1.0.5",
   "private": true,
+  "main": "src/index.js",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
